[v0.90.2][runtime] Continue splitting long_lived_agent.rs by responsibility (Closes #2310)
